requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
scipy
statsmodels
geopandas
shapely
rasterio
xarray
rioxarray
pyproj
folium
plotly
jupyter
pyarrow
pyyaml
